+++ pkg/analysis_server/lib/src/lsp/handlers/handler_document_color_presentation.dart
@@ -249,7 +249,7 @@ class DocumentColorPresentationHandler
           parent is PrefixedIdentifier ? parent.element : node.element;
 
       return switch (element) {
-        PropertyAccessorElement(:var variable) => variable?.isConst ?? false,
+        PropertyAccessorElement(:var variable) => variable.isConst,
         VariableElement() => element.isConst,
         _ => false,
       };