+++ pkg/analysis_server/lib/src/services/correction/util.dart
@@ -5,6 +5,7 @@
 import 'dart:math';
 
 import 'package:_fe_analyzer_shared/src/scanner/token.dart';
+import 'package:analysis_server/src/services/completion/dart/feature_computer.dart';
 import 'package:analyzer/dart/ast/precedence.dart';
 import 'package:analyzer/dart/ast/visitor.dart';
 import 'package:analyzer/dart/element/element.dart';
@@ -32,6 +33,26 @@ Expression climbPropertyAccess(Expression node) {
     }
     return node;
   }
+}
+
+/// Computes the [InterfaceElement] of the context type of a given
+/// [dotShorthandNode].
+///
+/// Returns `null` if no context type exists or the context type isn't an
+/// [InterfaceType].
+InterfaceElement? computeDotShorthandContextTypeElement(
+  AstNode dotShorthandNode,
+  LibraryElement libraryElement,
+) {
+  var featureComputer = FeatureComputer(
+    libraryElement.typeSystem,
+    libraryElement.typeProvider,
+  );
+  var contextType = featureComputer.computeContextType(
+    dotShorthandNode,
+    dotShorthandNode.offset,
+  );
+  return contextType is InterfaceType ? contextType.element : null;
 }
 
 /// Return references to the [element] inside the [root] node.