+++ pkg/analyzer/lib/src/fine/manifest_ast.dart
@@ -405,7 +405,7 @@ class _ElementCollector extends ThrowingAstVisitor<void> {
     // evaluation we will access the corresponding constant variable for
     // its initializer. So, we also depend on the variable.
     if (element is GetterElementImpl) {
-      _addElement(element.variable!);
+      _addElement(element.variable);
     }
   }
 }