+++ pkg/analyzer/lib/src/services/top_level_declarations.dart
@@ -110,10 +110,7 @@ class TopLevelDeclarations {
         libraryElement.exportNamespace.get2(baseName) ??
         libraryElement.exportNamespace.get2('$baseName=');
     if (element is PropertyAccessorElement) {
-      var variable = element.variable;
-      if (variable != null) {
-        return variable;
-      }
+      return element.variable;
     }
     return element;
   }