+++ pkg/analyzer/tool/messages/generate_test.dart
@@ -6,16 +6,15 @@
 // date.
 
 import 'package:analyzer_testing/package_root.dart' as pkg_root;
-import 'package:analyzer_utilities/tools.dart';
+import 'package:analyzer_utilities/generated_content_check.dart';
 import 'package:path/path.dart';
 
 import 'error_code_info.dart';
 import 'generate.dart';
 
 Future<void> main() async {
-  await GeneratedContent.checkAll(
+  await allTargets.check(
     pkg_root.packageRoot,
     join(analyzerPkgPath, 'tool', 'messages', 'generate.dart'),
-    allTargets,
   );
 }