+++ pkg/linter/lib/src/rules/exhaustive_cases.dart
@@ -100,10 +100,7 @@ extension on Element? {
   Element? get variableElement {
     var self = this;
     if (self is GetterElement) {
-      var variable = self.variable;
-      if (variable != null) {
-        return variable;
-      }
+      return self.variable;
     }
     return self;
   }